refactor(emailer): define functions inside object
- this is done to allow stubbing with sinon, since sinon only
accepts objects.
